+++ src/drawgfx.c
@@ -31,7 +31,7 @@ static UINT8 is_raw[TRANSPARENCY_MODES];
 
 INLINE UINT32 read_dword(void *address)
 {
-	if ((long)address & 3)
+	if ((size_t)address & 3)
 	{
   		return	(*((UINT8 *)address  ) << SHIFT0) +
 				(*((UINT8 *)address+1) << SHIFT1) +
@@ -45,7 +45,7 @@ INLINE UINT32 read_dword(void *address)
 
 INLINE void write_dword(void *address, UINT32 data)
 {
-  	if ((long)address & 3)
+  	if ((size_t)address & 3)
 	{
 		*((UINT8 *)address)   = (data>>SHIFT0);
 		*((UINT8 *)address+1) = (data>>SHIFT1);
@@ -301,7 +301,7 @@ INLINE void blockmove_NtoN_transpen_noremap8(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((size_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -369,7 +369,7 @@ INLINE void blockmove_NtoN_transpen_noremap_flipx8(
 	while (srcheight)
 	{
 		end = dstdata + srcwidth;
-		while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+		while (((size_t)srcdata & 3) && dstdata < end)	/* longword align */
 		{
 			int col;
 
@@ -3785,7 +3785,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_transpen,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata - dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata > end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata > end)	/* longword align */
 			{
 				int col;
 
@@ -3836,7 +3836,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_transpen,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata + dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata < end)	/* longword align */
 			{
 				int col;
 
@@ -3967,7 +3967,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_transblend,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata - dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata > end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata > end)	/* longword align */
 			{
 				int col;
 
@@ -4018,7 +4018,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_transblend,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata + dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata < end)	/* longword align */
 			{
 				int col;
 
@@ -4076,7 +4076,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_transmask,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata - dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata > end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata > end)	/* longword align */
 			{
 				int col;
 
@@ -4124,7 +4124,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_transmask,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata + dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata < end)	/* longword align */
 			{
 				int col;
 
@@ -4456,7 +4456,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_alphaone,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata - dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata > end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata > end)	/* longword align */
 			{
 				int col;
 
@@ -4544,7 +4544,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_alphaone,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata + dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata < end)	/* longword align */
 			{
 				int col;
 
@@ -4638,7 +4638,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_alpha,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata - dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata > end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata > end)	/* longword align */
 			{
 				int col;
 
@@ -4689,7 +4689,7 @@ DECLARE_SWAP_RAW_PRI(blockmove_8toN_alpha,(COMMON_ARGS,
 		while (dstheight)
 		{
 			end = dstdata + dstwidth*HMODULO;
-			while (((long)srcdata & 3) && dstdata < end)	/* longword align */
+			while (((size_t)srcdata & 3) && dstdata < end)	/* longword align */
 			{
 				int col;
